python-face-service/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
numpy
opencv-python
Pillow
deepface
tf-keras
